Game of Life 1.2
Fixed several bugs that carried over from the pre-Engine version. First,
when dragging the GUI Pane quickly and the Pane 'lags' behind the mouse,
living cells are no longer placed where the mouse drags. Second, the Pane
is no longer able to be dragged and stuck behind the Operation Bar in the
top-right corner, instead it 'slides' off the sides. Third, right-clicking
with a preset pattern selected removes the selected pattern. Finally, when
zooming in or out, the selected preset pattern is zoomed as well to fit
with the altered grid sizes.