Game of Life 1.3
Added several new features, most importantly dynamic pattern loading and
selection and Tooltips. All patterns are now saved within the patterns
folder and are loaded into the simulation at runtime. The pattern selector
has also been overhauled; the loaded patterns are now broken into folders
which can be expanded to reveal the individual patterns, which can be
selected and used from there. Second, Tooltips now appear when some
buttons and features are hovered over with the pointer for a certain
amount of time.