Game of Life 2.0
The Game of Life 2.0 brings a dynamically expanding rather than static
Grid, enhanced zooming, an expanded pattern library, and hotkeys. The Grid
will now expand infinitely as the simulation expands. Moving and zooming
have also been improved. Zooming will now center around the cursor and the
Grid can be zoomed out much farther. New hotkeys are available for Toolbar
buttons as well as the addititon of different functionality when the shift
and control keys are held.