Game of Life 3.0
Complete rebuild of the old Game of Life, including large performance improvements and bug fixes.
The Game of Life now supports selections which can be manipulated in a number of ways.
Revamped icons to a GNOME-based theme.
Improved performance of simulation.
Load images and patterns from within the source to separate the JAR from external folders.
Other improvements and fixes.

Uses the latest version of the Engine.